changes scale factor struct type from uint8_t to uint16_t

+++ Core/Inc/spl06-007.h
@@ -88,7 +88,7 @@ typedef struct {
 	float compensatedTemperature;
 
 	/* Scale Factor */
-	uint8_t scaleFactor;
+	uint16_t scaleFactor;
 
 } SPL06_007;
